requirements.txt
fastapi
uvicorn[standard]
prometheus-client
pdfplumber
pandas
pyyaml
python-dateutil
scikit-learn
joblib